# TFLite: reduced duplicated calculation of exp in softmax.h (float)